consolidate util tests
